+++ Snow Wars/Renderer.cpp
@@ -8,16 +8,16 @@ void Renderer::Render(const vec2& position)
 	float frameWidth = m_sprite->GetSize().x / m_frameCount;
 
 	SDL_Rect image;
-	image.x = (position.x - frameWidth) + (frameWidth / 2);
-	image.y = position.y - (m_sprite->GetSize().y / 2);
-	image.w = frameWidth;
-	image.h = m_sprite->GetSize().y;
+	image.x = (int)((position.x - frameWidth) + (frameWidth / 2));
+	image.y = (int)(position.y - (m_sprite->GetSize().y / 2));
+	image.w = (int)frameWidth;
+	image.h = (int)m_sprite->GetSize().y;
 
 	SDL_Rect frame;
-	frame.x = frameWidth * m_currentFrame;
+	frame.x = (int)(frameWidth * m_currentFrame);
 	frame.y = 0;
-	frame.w = frameWidth;
-	frame.h = m_sprite->GetSize().y;
+	frame.w = (int)frameWidth;
+	frame.h = (int)m_sprite->GetSize().y;
 	SDL_RenderCopy(Window::m_renderer, m_sprite->GetTexture(), &frame, &image);
 }
 
@@ -26,20 +26,20 @@ void Renderer::Render(const vec2& position, const float angle)
 	float frameWidth = m_sprite->GetSize().x / m_frameCount;
 
 	SDL_Rect image;
-	image.x = (position.x - frameWidth) + (frameWidth / 2);
-	image.y = position.y - (m_sprite->GetSize().y / 2);
-	image.w = frameWidth;
-	image.h = m_sprite->GetSize().y;
+	image.x = (int)((position.x - frameWidth) + (frameWidth / 2));
+	image.y = (int)(position.y - (m_sprite->GetSize().y / 2));
+	image.w = (int)frameWidth;
+	image.h = (int)m_sprite->GetSize().y;
 
 	SDL_Rect frame;
-	frame.x = frameWidth * m_currentFrame;
+	frame.x = (int)(frameWidth * m_currentFrame);
 	frame.y = 0;
-	frame.w = frameWidth;
-	frame.h = m_sprite->GetSize().y;
+	frame.w = (int)frameWidth;
+	frame.h = (int)m_sprite->GetSize().y;
 
 	SDL_Point point;
-	point.x = frameWidth - (frameWidth / 2);
-	point.y = m_sprite->GetSize().y / 2;
+	point.x = (int)(frameWidth - (frameWidth / 2));
+	point.y = (int)(m_sprite->GetSize().y / 2);
 
 	SDL_RenderCopyEx(Window::m_renderer, m_sprite->GetTexture(), &frame, &image, angle, &point, SDL_FLIP_NONE);
 }
